+++ speed/adolc/ode.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-or-later
 // SPDX-FileCopyrightText: Bradley M. Bell <bradbell@seanet.com>
-// SPDX-FileContributor: 2003-22 Bradley M. Bell
+// SPDX-FileContributor: 2003-24 Bradley M. Bell
 // ----------------------------------------------------------------------------
 /*
 {xrst_begin adolc_ode.cpp}
@@ -51,7 +51,7 @@ bool link_ode(
    typedef CppAD::vector<double>  DblVector;
 
    size_t i, j;
-   int tag    = 0;       // tape identifier
+   short tag    = 0;       // tape identifier
    int keep   = 0;       // do not keep forward mode results
    size_t p   = 0;       // use ode to calculate function values
    size_t n   = size;    // number of independent variables
@@ -65,11 +65,11 @@ bool link_ode(
    size_t size_out;           // capacity of an allocation
 
    // raw memory for use with adolc
-   size_min = n;
+   size_min = size_t(n);
    double *x_raw   = thread_alloc::create_array<double>(size_min, size_out);
-   size_min = m * n;
+   size_min = size_t(m * n);
    double *jac_raw = thread_alloc::create_array<double>(size_min, size_out);
-   size_min = m;
+   size_min = size_t(m);
    double **jac_ptr = thread_alloc::create_array<double*>(size_min, size_out);
    for(i = 0; i < m; i++)
       jac_ptr[i] = jac_raw + i * n;
@@ -77,7 +77,7 @@ bool link_ode(
    // -------------------------------------------------------------
    if( ! global_option["onetape"] ) while(repeat--)
    {  // choose next x value
-      uniform_01(n, x);
+      uniform_01( size_t(n), x);
 
       // declare independent variables
       trace_on(tag, keep);
@@ -95,11 +95,11 @@ bool link_ode(
       // evaluate the Jacobian
       for(j = 0; j < n; j++)
          x_raw[j] = x[j];
-      jacobian(tag, m, n, x_raw, jac_ptr);
+      jacobian(tag, int(m), int(n), x_raw, jac_ptr);
    }
    else
    {  // choose next x value
-      uniform_01(n, x);
+      uniform_01( size_t(n), x);
 
       // declare independent variables
       trace_on(tag, keep);
@@ -116,12 +116,12 @@ bool link_ode(
 
       while(repeat--)
       {  // get next argument value
-         uniform_01(n, x);
+         uniform_01( size_t(n), x);
          for(j = 0; j < n; j++)
             x_raw[j] = x[j];
 
          // evaluate jacobian
-         jacobian(tag, m, n, x_raw, jac_ptr);
+         jacobian(tag, int(m), int(n), x_raw, jac_ptr);
       }
    }
    // convert return value to a simple vector
